Linux-specific fix for the provider test
[#184673864](https://www.pivotaltracker.com/story/show/184673864)
